Move TaskController into its own component file
